+++ kernel/Arch/Arch.hpp
@@ -10,7 +10,7 @@
  */
 #pragma once
 
-#include <meta/compiler.hpp>
+#include <Support/compiler.hpp>
 #include <stddef.h>
 #include <stdint.h>
 #if defined(__i686__)
