+++ kernel/Devices/graphics/framebuffer.hpp
@@ -10,7 +10,7 @@
  */
 #pragma once
 #include <stdint.h>
-#include <meta/compiler.hpp>
+#include <Support/compiler.hpp>
 
 namespace Graphics {
