+++ kernel/Library/Bitset.hpp
@@ -14,7 +14,7 @@
 #include <Arch/Arch.hpp>
 #include <Library/string.hpp>
 #include <limits.h>
-#include <meta/compiler.hpp>
+#include <Support/compiler.hpp>
 #include <stddef.h>
 #include <stdint.h>
