+++ kernel/Memory/MemorySection.hpp
@@ -10,7 +10,7 @@
  */
 #pragma once
 #include <Arch/Memory.hpp>
-#include <meta/compiler.hpp>
+#include <Support/compiler.hpp>
 #include <stddef.h>
 #include <stdint.h>
