+++ kernel/Memory/paging.cpp
@@ -18,7 +18,7 @@
 #include <Library/string.hpp>
 #include <Memory/Physical.hpp>
 #include <Memory/paging.hpp>
-#include <meta/sections.hpp>
+#include <Support/sections.hpp>
 #include <sys/Panic.hpp>
 #include <stddef.h>
