+++ kernel/Support/gcc.cpp
