+++ kernel/sys/Panic.hpp
@@ -10,7 +10,7 @@
  */
 #pragma once
 
-#include <meta/compiler.hpp>
+#include <Support/compiler.hpp>
 #include <Arch/Arch.hpp>
 
 /**
